refactor: remove constraint collector int overloads (#1039)

+++ java/order-picking/src/main/java/org/acme/orderpicking/solver/OrderPickingConstraintProvider.java
@@ -101,7 +101,7 @@ public class OrderPickingConstraintProvider implements ConstraintProvider {
                 .asConstraint("Minimize the distance from last trolley pick to the path origin");
     }
 
-    private int calculateOrderRequiredBuckets(int orderVolume, int bucketVolume) {
+    private long calculateOrderRequiredBuckets(long orderVolume, long bucketVolume) {
         return (orderVolume + (bucketVolume - 1)) / bucketVolume;
     }
 }
